+++ src/app/txid/[txId]/redesign/TenureChangePage.tsx
@@ -8,7 +8,6 @@ import {
 import { TenureAlert } from './Alert';
 import { TxHeader } from './TxHeader';
 import { TxSummary } from './TxSummary';
-import { TxSummaryTable } from './TxSummaryTable';
 
 export const TenureChangePage = ({
   tx,
@@ -21,7 +20,7 @@ export const TenureChangePage = ({
         <TxHeader tx={tx} />
         <TenureAlert />
       </Stack>
-      <TxSummaryTable tx={tx} />
+      <TxSummary tx={tx} />
     </>
   );
 };
