+++ src/app/txid/[txId]/redesign/TxSummary.tsx
@@ -6,10 +6,11 @@ import { capitalize } from '@/common/utils/utils';
 import { Badge, BlockHeightBadge, DefaultBadgeLabel, TransactionStatusBadge } from '@/ui/Badge';
 import { Text } from '@/ui/Text';
 import StacksIconThin from '@/ui/icons/StacksIconThin';
-import { Flex, Grid, Icon, Stack, Table } from '@chakra-ui/react';
-import React from 'react';
+import { Flex, Icon, Stack, Table } from '@chakra-ui/react';
 
 import {
+  CoinbaseTransaction,
+  MempoolCoinbaseTransaction,
   MempoolTenureChangeTransaction,
   MempoolTokenTransferTransaction,
   MempoolTransaction,
@@ -20,7 +21,7 @@ import {
 
 import { useTxIdPageData } from '../TxIdPageContext';
 
-function SummaryItemLabel({ label }: { label: string }) {
+export function SummaryItemLabel({ label }: { label: string }) {
   return (
     <Text textStyle="text-medium-sm" color="textSecondary">
       {label}
@@ -28,7 +29,7 @@ function SummaryItemLabel({ label }: { label: string }) {
   );
 }
 
-function SummaryItemValue({
+export function SummaryItemValue({
   value,
   valueRenderer,
   label,
@@ -66,7 +67,7 @@ function SummaryItemValue({
   );
 }
 
-function PriceSummaryItemValue({ value }: { value: string }) {
+export function PriceSummaryItemValue({ value }: { value: string }) {
   const { stxPrice } = useTxIdPageData();
   const usdValue = stxPrice * Number(value);
 
@@ -291,7 +292,7 @@ export const TenureChangeTxSummaryItems = ({
         label="From"
         value={tx.sender_address}
         valueRenderer={value => (
-          <AddressLink principal={value} wordBreak="break-all">
+          <AddressLink principal={value} wordBreak="break-all" variant="tableLink">
             {value}
           </AddressLink>
         )}
@@ -384,23 +385,112 @@ export const TenureChangeTxSummaryItems = ({
   );
 };
 
+export const CoinbaseTxSummaryItems = ({
+  tx,
+}: {
+  tx: CoinbaseTransaction | MempoolCoinbaseTransaction;
+}) => {
+  console.log('coinbase tx', { tx });
+  return (
+    <>
+      <SummaryItem label="ID" value={tx.tx_id} copyable />
+      <SummaryItem
+        label="Status"
+        value={tx.tx_status}
+        valueRenderer={value => <TransactionStatusBadge tx={tx} />}
+        copyable
+      />
+      <SummaryItem
+        label="From"
+        value={tx.sender_address}
+        valueRenderer={value => (
+          <AddressLink principal={value} wordBreak="break-all" variant="tableLink">
+            {value}
+          </AddressLink>
+        )}
+        copyable
+      />
+      {tx.coinbase_payload?.alt_recipient && (
+        <SummaryItem
+          label="Reward Recipient"
+          value={tx.coinbase_payload?.alt_recipient}
+          valueRenderer={value => (
+            <AddressLink principal={value} wordBreak="break-all" variant="tableLink">
+              {value}
+            </AddressLink>
+          )}
+          copyable
+        />
+      )}
+      {'block_time' in tx && tx.block_time && (
+        <SummaryItem
+          label="Timestamp"
+          value={formatBlockTime(tx.block_time)}
+          valueRenderer={value => (
+            <Badge variant="solid">
+              <DefaultBadgeLabel label={value} />
+            </Badge>
+          )}
+          copyable
+        />
+      )}
+      <SummaryItem
+        label="Fee"
+        value={tx.fee_rate}
+        valueRenderer={value => <PriceSummaryItemValue value={value} />}
+      />
+      <SummaryItem label="Nonce" value={tx.nonce?.toString() || ''} copyable />
+      {'block_height' in tx && tx.block_height && (
+        <SummaryItem
+          label="Block height"
+          value={tx.block_height?.toString() || ''}
+          copyable
+          valueRenderer={value => <BlockHeightBadge blockType="stx" blockHeight={Number(value)} />}
+        />
+      )}
+      {'block_hash' in tx && tx.block_hash && (
+        <SummaryItem
+          label="Block hash"
+          value={tx.block_hash?.toString() || ''}
+          copyable
+          valueRenderer={value => (
+            <BlockLink hash={value} wordBreak="break-all">
+              {value}
+            </BlockLink>
+          )}
+        />
+      )}
+      {/* <SummaryItem
+        label="Tenure height"
+        value={tx.tenure_change_payload?.previous_tenure_end || ''}
+        copyable
+        valueRenderer={value => <BlockHeightBadge blockType="stx" blockHeight={Number(value)} />}
+      /> */}
+      {'burn_block_height' in tx && tx.burn_block_height && (
+        <SummaryItem
+          label="Bitcoin Anchor"
+          value={tx.burn_block_height?.toString() || ''}
+          copyable
+          valueRenderer={value => <BlockHeightBadge blockType="btc" blockHeight={Number(value)} />}
+        />
+      )}
+    </>
+  );
+};
+
 export function TxSummary({ tx }: { tx: Transaction | MempoolTransaction }) {
-  let summary;
-  if (tx.tx_type === 'coinbase') summary = null;
-  if (tx.tx_type === 'token_transfer') summary = <TokenTransferTxSummaryItems tx={tx} />;
-  if (tx.tx_type === 'contract_call') summary = null;
-  if (tx.tx_type === 'smart_contract') summary = null;
-  if (tx.tx_type === 'tenure_change') summary = <TenureChangeTxSummaryItems tx={tx} />;
+  let summaryContent;
+  if (tx.tx_type === 'coinbase') summaryContent = <CoinbaseTxSummaryItems tx={tx} />;
+  if (tx.tx_type === 'token_transfer') summaryContent = <TokenTransferTxSummaryItems tx={tx} />;
+  if (tx.tx_type === 'contract_call') summaryContent = null;
+  if (tx.tx_type === 'smart_contract') summaryContent = null;
+  if (tx.tx_type === 'tenure_change') summaryContent = <TenureChangeTxSummaryItems tx={tx} />;
 
   return (
-    <Flex borderRadius="redesign.xl" border="1px solid" borderColor="redesignBorderSecondary" p={3}>
-      <Grid
-        templateColumns={{ base: '1fr', md: 'minmax(auto, max-content) 1fr' }}
-        gap={{ base: 6, md: 0 }}
-        // columnGap={6}
-      >
-        {summary}
-      </Grid>
-    </Flex>
+    <Box borderRadius="redesign.xl" border="1px solid" borderColor="redesignBorderSecondary" p={3}>
+      <Table.Root w="full">
+        <Table.Body>{summaryContent}</Table.Body>
+      </Table.Root>
+    </Box>
   );
 }
