+++ src/common/components/CopyButton.tsx
@@ -54,43 +54,3 @@ export const CopyButtonRedesign = ({
     </Button>
   );
 };
-
-export const CopyButton2 = ({
-  initialValue,
-  buttonProps,
-  iconProps,
-}: {
-  initialValue: string;
-  iconProps: IconProps;
-  buttonProps: ButtonProps;
-}) => {
-  const { copied, value, setValue, copy } = useClipboard({
-    value: initialValue,
-    timeout: 750,
-  });
-  return (
-    <Button
-      onClick={() => copy()}
-      variant="unstyled"
-      {...buttonProps}
-      bg={copied ? 'surfaceInvert' : 'transparent'}
-      _hover={{ bg: copied ? 'surfaceInvert' : 'surfaceFifth' }}
-      _groupHover={{ bg: copied ? 'surfaceInvert' : 'surfaceTertiary' }}
-      borderRadius="redesign.sm"
-      minWidth="0"
-      h="fit-content"
-      w="fit-content"
-      display="flex"
-      alignItems="center"
-      justifyContent="center"
-    >
-      <Icon
-        {...iconProps}
-        color={copied ? 'iconInvert' : 'iconSecondary'}
-        _hover={{ color: copied ? 'iconInvert' : 'iconPrimary' }}
-      >
-        {copied ? <Check /> : <CopySimple />}
-      </Icon>
-    </Button>
-  );
-};
